requirements.txt
Flask==3.1.2
Werkzeug==3.1.4
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-Mail==0.10.0
bcrypt==4.2.0
python-dotenv==1.0.1
email-validator==2.2.0
pytest==8.3.3
pytest-flask==1.3.0
requests==2.31.0
PyMySQL==1.1.1
cryptography==42.0.5
PyJWT==2.8.0

